refactoring pseudobam
