+++ src/main.cpp
@@ -528,6 +528,10 @@ bool CheckOptionsEM(ProgramOptions& opt, bool emonly = false) {
       cerr << "Warning: you asked for " << opt.threads
            << ", but only " << n << " cores on the machine" << endl;
     }
+    if (opt.threads > 1 && opt.pseudobam) {
+      cerr << "Error: pseudobam is not compatible with running on many threads."<< endl;
+      ret = false;
+    }
   }
 
   if (opt.bootstrap < 0) {
